backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
httpx==0.25.2
redis==5.0.1
neo4j==5.15.0
psycopg2-binary==2.9.9
